Refactor auth and payment routes into modular routers
